+++ src/components/CountIcon/styles.js
@@ -2,20 +2,20 @@ import { StyleSheet } from 'react-native';
 
 import styleConstants from '../../styleConstants';
 
+const SIZE = 20;
+
 const styles = StyleSheet.create({
   container: {
-    width: 20,
-    height: 20,
-    borderRadius: 10,
-    overflow: 'hidden', // ios
-    backgroundColor: styleConstants.dividerColor,
+    width: SIZE,
+    height: SIZE,
+    borderRadius: SIZE / 2,
+    backgroundColor: styleConstants.colors.dividerColor,
     justifyContent: 'center',
     alignItems: 'center',
   },
   text: {
-    fontSize: styleConstants.verySmallFont,
-    color: styleConstants.primaryText,
-    marginBottom: 1.5,
+    fontSize: styleConstants.fonts.sizes.verySmall,
+    color: styleConstants.colors.primaryText,
   },
 });